Forum ok avec ajout de topic et de post

+++ src/Controller/PostController.php
@@ -3,10 +3,15 @@
 namespace App\Controller;
 
 use App\Entity\Post;
+use App\Entity\Topic;
+use App\Form\PostType;
+use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\Persistence\ManagerRegistry;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+
 
 class PostController extends AbstractController
 {
@@ -20,36 +25,82 @@ class PostController extends AbstractController
         ]);
     }
 
-    #[Route('/post/create', name: 'add_post')]
-    public function add(Request $request): Response
+    #[Route('/post/topic/{id}', name: 'app_topic_show')]
+    public function show(Request $request, Topic $topic, ManagerRegistry $doctrine): Response
     {
+        $posts = $doctrine->getRepository(Post::class)->findBy(['topic' => $topic]);
+    
         $post = new Post();
         $form = $this->createForm(PostType::class, $post);
-
+    
         $form->handleRequest($request);
-
+    
         if ($form->isSubmitted() && $form->isValid()) {
+            $post->setTopic($topic);
+            $post->setUser($this->getUser());
+            $post->setDatePost(new \DateTime());
+    
+            $entityManager = $doctrine->getManager();
+            $entityManager->persist($post);
+            $entityManager->flush();
+    
+            // Redirection vers la page d'affichage du topic
+            return $this->redirectToRoute('app_topic_show', ['id' => $topic->getId()]);
+        }
+    
+        return $this->render('topic/show.html.twig', [
+            'topic' => $topic,
+            'posts' => $posts, // Assurez-vous que cette variable est définie et transmise au template
+            'form' => $form->createView(),
+        ]);
+    }
+    
+
+
+    #[Route('/post/create/{id}', name: 'add_post')]
+    public function add(Request $request, $id): Response
+    {
+        $topic = $this->getDoctrine()->getRepository(Topic::class)->find($id);
+    
+        if (!$topic) {
+            throw $this->createNotFoundException('Le topic n\'existe pas.');
+        }
+    
+        $post = new Post();
+        $form = $this->createForm(PostType::class, $post);
+    
+        $form->handleRequest($request);
+    
+        if ($form->isSubmitted() && $form->isValid()) {
+            $post->setTopic($topic);
+            $post->setUser($this->getUser());
+            $post->setDatePost(new \DateTime());
+    
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->persist($post);
             $entityManager->flush();
-
-            // Redirection vers la page d'affichage du post nouvellement créé
-            return $this->redirectToRoute('app_post_show', ['id' => $post->getId()]);
+    
+            // Redirection vers la page d'affichage du topic
+            return $this->redirectToRoute('app_topic_show', ['id' => $topic->getId()]);
         }
-
+    
         return $this->render('post/add.html.twig', [
             'form' => $form->createView(),
         ]);
     }
 
     #[Route('/post/{id}/delete', name: 'delete_post')]
-    public function delete(ManagerRegistry $doctrine, Post $post): Response
+    public function deletePost(EntityManagerInterface $entityManager, Post $post): Response
     {
-        $entityManager = $doctrine->getManager();
+        // Vérifier si l'utilisateur actuel est l'auteur du post
+        if ($post->getUser() !== $this->getUser()) {
+            throw new AccessDeniedException('Vous n\'êtes pas autorisé à supprimer ce post.');
+        }
+    
         $entityManager->remove($post);
         $entityManager->flush();
-
-        return $this->redirectToRoute('app_categorie');
-
+    
+        // Redirection vers la page d'affichage du topic
+        return $this->redirectToRoute('app_topic_show', ['id' => $post->getTopic()->getId()]);
     }
 }
